Add Menu Slide, add HKMoreViewController.

+++ HKMine/HKBoardView.h
@@ -20,8 +20,7 @@
 
 @interface HKBoardView : UIView <UIGestureRecognizerDelegate>
 
-//@property (nonatomic, assign) NSUInteger rowCount, columnCount, markedNumber;
-@property (nonatomic, assign) int rowCount, columnCount, markedNumber;
+@property (nonatomic, assign) NSUInteger rowCount, columnCount, markedNumber;
 
 @property (nonatomic, assign) CGFloat sideLength;
 @property (nonatomic, strong) NSMutableArray *cellsStates;
